refactor(opencode): organize message event modules
